+++ CircleCollider.hpp
@@ -14,18 +14,18 @@ namespace spic {
     class CircleCollider : public Collider {
         public:
 
-            CircleCollider(double radius, std::vector<int> collideLayers, Vector2<float> offset) : Collider(collideLayers, offset);
+            CircleCollider(float radius, std::vector<int> collideLayers, Vector2<float> offset) : Collider(collideLayers, offset);
             /**
              * @brief Get the collider's radius
              * @return The radius
              */
-            double Radius() const;
+            float Radius() const;
 
             /**
              * @brief Set the collider's radius
              * @param newRadius The desired radius
              */
-            void Radius(double newRadius);
+            void Radius(float newRadius);
 
         private:
         #include "CircleCollider_private.hpp"
