+++ Transform.hpp
@@ -10,8 +10,8 @@ namespace spic {
      */
     class Transform {
     public:
-        Transform(const Vector2<double>& pos, double rot, double scale);
-        void LookAt(const Vector2<double>& pos);
+        Transform(const Vector2<float>& pos, float rot, float scale);
+        void LookAt(const Vector2<float>& pos);
     private:
     #include "Transform_private.hpp"
     };
